Create windows.lib with LLVM tools rather than MSVC
As mentioned in the previous commit, the libraries produced by
llvm-dlltool/llvm-ar are not practically significantly different from
the ones produced with MSVC. The lack of the lexically ordered linker
member is not preventing MSVC's link.exe from working.

On the other hand, using the LLVM tools is:
- faster
- makes the .lib files 25% smaller.
- is more deterministic (no timestamps in the resulting files)
- allows to build the libraries on any OS that has the LLVM tools, not
  just Windows.

This change uses the code that used to be under crates/tools/gnullvm,
changes the use of where.exe for cross-platform support, and adds
support for i686.